refactor: Merge in Splitter too
