+++ src/components/controllers/repl/index.jsx
@@ -1,10 +1,9 @@
 import { useState, useEffect } from 'preact/hooks';
 import { useLocation, useRoute, ErrorBoundary } from 'preact-iso';
-import { Splitter } from '../../splitter';
 import { textToBase64 } from './query-encode.js';
 import { EXAMPLES, fetchExample } from './examples';
 import { useStoredValue } from '../../../lib/localstorage';
-import { CodeEditor, Runner, ErrorOverlay } from '../../routes';
+import { CodeEditor, Runner, ErrorOverlay, Splitter } from '../../routes';
 import { parseStackTrace } from './errors';
 import style from './style.module.css';
 import REPL_CSS from './examples/style.css?raw';
