+++ src/components/controllers/tutorial/index.jsx
@@ -12,9 +12,8 @@ import { ErrorBoundary, useLocation } from 'preact-iso';
 import { TutorialContext, SolutionContext } from './contexts';
 import { parseStackTrace } from '../repl/errors';
 import cx from '../../../lib/cx';
-import { CodeEditor, Runner, ErrorOverlay } from '../../routes';
+import { CodeEditor, Runner, ErrorOverlay, Splitter } from '../../routes';
 import { useLanguage } from '../../../lib/i18n';
-import { Splitter } from '../../splitter';
 import config from '../../../config.json';
 import { MarkdownRegion } from '../markdown-region';
 import style from './style.module.css';
