+++ src/lib/repl.js
@@ -1,3 +1,4 @@
 export { default as CodeEditor } from '../components/code-editor';
 export { default as Runner } from '../components/controllers/repl/runner';
 export { ErrorOverlay } from '../components/controllers/repl/error-overlay';
+export { Splitter } from '../components/splitter';
